requirements.txt
Flask
RPi.GPIO
gpiozero